+++ tests/e2e/cases/dynamic_require_static_buffer_equals/beq_true.ts
@@ -1,0 +1,1 @@
+module.exports.label = "beq_true";